all the logger noise from GLiNER and GLiREL is pathetic
